4-5-2: Added new Makefile recipe
I added a new Makefile recipe for the work being done on this branch.  452 currently compiles the Map_Memory library, then compiles the MapMemTest.c and finally links that code against the libarary.  In the near future, it will also compile Elf_Manipulation in addition to the 4-5 main(), which I have yet to begin writing.