4-5-2: 'Main' Initial Commit
'Main' is actually 4-5_DCE_Practice-2_ModEntryPoint.c and 4-5_DCE_Practice-2_ModEntryPoint.c is the 'glue' that will utilize both Elf_Manipulation.h and Map_Memory.h to make changes to an ELF file.  Currently, it only implements functionality that is, well, implemented.  It maps a file into memory, verifies it's an elf, and then cleans up.  Next?  Implement functionality into Elf_Manipulation.h that begins to fill in information about the file in a SUPER-struct.