4-5-2: Map_Mem.h MACRO update
Added MM_TYPE_CAVE macro to the "memory type" list.  Since I already have a framework surrounding memory references and their sizes, I also want to use this framework for when I start looking for "code caves".  MM_TYPE_CAVE will indicate to the free_struct() function that the fileMem_ptr isn't owned by that struct and that it should not be free()'d.  Hell, it's likely not even a pointer to the beginning of allocated space.  free_struct() should just zero the member variables and that is all.  Now, on to modify free_struct()...