4-5-2: Implemented M_M.h validate_struct()
Instead of performing input validation in a million different functions, I wrote a libarary function (validate_struct) into Map_Memory.h to do the work for me.  New additions to the struct and/or the memory type macros would constitute a change to the algorithm anyway so it made sense to extricate it early on.