3-10-2: Fixed Seg Fault in main()
I assumed the Seg Fault was coming from the free_struct*() functions since that process can be pretty complicated and easily corrupt memory.  Not this time.  The problem arose because I wasn't properly controlling a while loop in main() as it 'pointer math'd it's way down the array of struct pointers.  I wasn't properly looking for the NULL terminator at the end of the array and it was sliding off the end.  Whoops.  All better.