3-10-2: Improved Hp.c free_PID_struct_arr()
I didn't like the way free_PID_struct_arr() was free()ing the struct pointers.  I was using a struct_ptr variable to aid in the process but, as it turns out, it works better (cleaner) when I just use the array pointer itself (since the free_struct*() functions want the pointer to the thing so they can 'NULL' the variable as well.  It's all nice and pretty now.  I commented out the DEBUGGING code and the struct_ptr but didn't delete it.  (Commented but not forgotten)