3-10-2: The Harkledir refactor has begun
I decided I wanted to refactor the directoryDetails struct.  This version will essentially before Harkledir v2 so I added a Changelog(?) to the top of the header file.

I wanted a better way to store information gleaned from dirent structs within dirDetails.  I decided on a new harkledirEnt struct.  Not only will all of the dirDetails support functions need to be updated, I'll need to add new functionality for the new harkledirEnt struct.  This is the beginning.

Currently, the only code I've changed is the format of dirDetails and adding harkleDirEnt.  I haven't even prototyped any of the new harkleDirEnt functions yet.