3-10-2: BUG HUNT OVER (Harkledir.c)
The ultimate problem turned out to be I wasn't allocating the right 'chunks' of memory in the dirDetails array members.  I spent all day refactoring all of these headers to utilize the new structs and the new struct format.  I missed the realloc for the array sizes.  That realloc statement was still a holdover from when I was allocating space for char*, not struct pointers.  Now that I think about it, I may have been inadavertently misalinging the stack (if that's even possible in C).

Regardless, the problem was discovered in my free() functions because (and I'm guessing here cuz I'm real fatigued) there was bleedover into non-allocated memory space.  Regardless, I'm back where I was at (basically) prior to the huge refactor.  The program lists the PIDs, associates the PIDs with their cmdline arguments (if any), prompts the user for a PID to examine, translates that PID number into an absolute path.  Wheeeeeeeeeeeee...

As I type this, I'm thinking of a couple more things to fix/tweak in Harkledir.c.