3-10-2: BUG HUNT size_a_file() needs abs path
I'm in the middle of a bug hunt but I'm transitioning from my laptop to my desktop.  The situation is that the symlink's aren't getting sized.  size_a_file() is failing because it's not getting an absolute path to the file being sized.  If memory serves, I was working on temporarily building an abs path for the file in question (either in size_a_file() or the calling function) but I don't recall where I stopped.