3-10-2: Added Hello World
I know I have Hello Worlds elsewhere in this project but I wanted one here with a 'pause' in the middle.  It gave me a PID that was very basic to look and also happened to be owned by me.  If I couldn't properly parse details of a HelloWorld that I executed, how can I expect to parse the world?!