3-10-2: BUG HUNT Hd.c readlink()
I was perplexed why readlink() wasn't resolving symbolic links.  Up until now, I blamed it on /proc.  It wasn't until I tried it on a non-proc directory that I realized I was doing something wrong.  Silly me... I was passing in a relative path instead of an absolute path for the file to readlink() from.  Fixed.