3-10-2: print_PID_Libraries
This source code takes advantage of all the header files and examples that have gone before it.  It is simple in its execution.

usage:  parse_PID_libraries <PID>
output: A unique list of all the files mapped into memory by this process

There are still some bugs and this code could use some clean up but it's good enough, as is, that it will probably get merged into "practice" real soon.