3-10-2: MEMORY BUG HUNT free_char_arr() Mr.c FIXED
Only one char* in the char* array was being free()d because of some faulty copy/paste and/or auto-completion in my editor.  I was advancing to the next character in a memset, free()d string instead of advancing to the next char* in the char**.  FIXED.  Tested in valgrind.  (still other errors though)