3-4-1: nosig.c Initial Commit
It's mostly written as is.  There's a couple of /// IMPLEMENT LATER /// sections but I bet it would work well for a binary that doesn't require arguments (e.g., HelloWorld.exe).  After I get the fork() and execv() calls working, then it's time to write a signal handler.  The concept of "nosig" seems relatively easy.  I hope I'm not being deceived.  Not compiled.  Not tested.  Not done.  Written in a text editor.  Last stop?  execv()-town.