3-4-1: strace output from rt_sigaction BUG HUNT
I'm deciding to permanently(?) keep the strace output I used to help deconflict sigaction function calls against rt_sigaction system calls.  The information was very illuminating and it led me to make some interesting observations.  It also started me on a research journey that will, if not already, be documented in "Research Documents".  The 'base' notes currently exist as a comment in the as-of-yet un-committed nosig.c.  These files names are referenced in the Research Document named 3-4-1-rt_sigaction.txt.