3-18-1: BUG HUNT in "main.c"'s call to p_e_p()
BUG HUNT in 3-18_Threading_Practice-1_Grand_Prix.c's function call to plot_ellipse_points().  plot_ellipse_points() was being called with the dimensions of the window as the major and minor axis for the ellipse.  Now, it's being passing in half of that with care to round down in appropriate ways.  Written offline.  Untested.  Unknown if this fixes the bug.