3-18-1: Added member to Harklecurse.h hcCartCoord
Added a new member to the Harklecurse.h hcCartCoord struct.
    char default;  // Original character to print at this coordinate
This new member will hold the original character to print here.  It occurred to me that if "char graphic" is ever updated with a "racer token", the program won't necessarily remember the character that was originally there.  This will be a big deal if I implement functionality to make the track look better with ASCII art (e.g., slashes at the curves).