3-18-1: Added spacing to "main.c"'s track window
Added a dynamically sized "internal buffer" to aid in plotting the ellipse points.  Currently, the ellipse plot is "overflowing" the ncurses track window and mvwaddch() is failing.  I added logic to calculate an internalBuffer to allow for visual spacing between the ellipse plot points and the border of the track window.  Written offline.  Untested.