3-18-1: Modified Harklepipe.* w/ pipe read/write()
Added read_a_pipe() and write_a_pipe() functions to Harklepipe.*.  Unknown if write_a_pipe() is working properly.  It's possible that read_a_pipe() is the cause of a bug I'm currently hunting down.  This bug has been commented on in the "main.c" but it has to do with read_a_pipe().

 Changes to be committed:
	modified:   Harklepipe.c
	modified:   Harklepipe.h