3-18-1: "main.c" is linking but non-functional
There is a bug I'm currently hunting down that appears to be related to read_a_pipe().  It's commented in 3-18_Threading_Practice-1_Grand_Prix.c.

"main.c" has had a lot of thrash lately as I've been implementing new functionality.  It's gotten a bit unwieldy in it's size but I don't see a clear-cut/easy-way to resolve that concern right now.  Maybe we'll worry about it after the track plots and the thread racers show up on the track.