3-18-1: Refactored read_a_pipe() in Harklepipe.*
Added a parameter in read_a_pipe() to return the errno received on a failed read.  In a O_NONBLOCKing situation, it's perfectly normal to get EAGAIN if there's nothing to read.  Now, read_a_pipe() doesn't bark about the errno it gets, it just returns it in a new [OUT] parameter called errNumber.  Let the calling function worry about the errno.  This new functionality appears to be working perfectly.