3-18-1: Rando header Initial Commit
I originally collaborated on the creation of this header on a different project.  I cleaned up the code and organization to bring it into alignment with the unofficial, undocumented, internalized style guide I'm using as part of this repo.  Regardless, I need it here because I want to start psuedo-randomizing values for the threadd racers.