3-18-1: Modified "main.c" to get real racer distro
Modified 3-18_Threading_Practice-1_Grand_Prix.c in an attempt to get some manner of distribution from the racer threads.  I changed out the sleepy threads for threads that randomize a number and determine if the number is prime.  If it's prime, they update the counter in the pipe to the main thread.

That wasn't even working so well.  All of the racer's were clogged up in the same track point.  Then, I increased the size of the random number from unsigned int to unsigned long to unsigned long long.  Even then, I wasn't getting good distribution (IMHO) so I added a sub counter and a macro to control its size.  Now, each racer thread has to randomize 10,000 prime unsigned long longs before it will report back to the main thread.

I can see some distribution but I need to work on it some.