3-18-1: Began implementing laps in "main.c"
Added "highLap" variable to store the lap number of the current font runner.
Modified the "win" criteria to include currLap == numLaps in addition to having completed all the positions
Added a code block to call highest_lap() to determine the lap number of the front runner.  highest_lap() is not yet written.

Written offline.  Untested.