3-18-1: Added Research Source to README
Sometimes, the compiler (or linker) will optimize out calls to memset() that are attempting to zeroize memory just before it is free()d.  Likely, this will be the first topic I explore when I start Reverse Engineering binaries.  I want to know exactly what and how the compiler is doing to my memset() statements in functions like free_a_thing().