3-22-1: injector.c EOD commit
3-22_Process_Injection-1_injector.c is coming along.  injector.exe can currently:
	- Parse args
	- Validate input
	- Parse /proc/
	- Attack to a PID
	- Backup that PID's registers
	- Parse /proc/PID/maps to find executable memory
	- Backup that memory locally

To Do:
	- 'Procure' code to inject
	- Overwrite the memory section
	- Update RIP to point to the injected code
	- Resume execution of the PID
	- Wait for a SIGTRAP
	- Restore the PID to its full original state

Everything that currently needs to be cleaned up is being cleaned up.  main() is pretty clean as-is.  So far, I've done a good job of exricating most of the functionality to files 'outside' of main.c.  There's only a single swath of DEBUGGING code (currently commented out) that I used to look at the "memory backup" because I didn't want to use gdb.