3-22-1: BUG HUNT injector.c
Still working on the "Bad Address" error... it's entirely possible that my specific Linux implementation (Ubuntu 16.04) won't allow a process' mapped memory to have rwxp permissions.  It's possible that it will only permit r-xp or rw-p at any given time.  To test that theory, I implemented code in Section 6 to add the following *new functionality:

    - Find r-xp memory
    - Backup the r-xp memory
    - Change the r-xp memory to rw-p permissions
    - Write to the memory location
    * Change the rw-p permissions back to r-xp
    - Point RIP at the executable code
blahblahblah

If this works, I'll need to use this technique again when I attempt to restore the original contents of the mapped memory block and then restore the original registers.  I'll need to:
    - Set it to rw-p
    - Restore the contents
    - Set it to r-xp
    - Restore the registers
    - Continue the PID on its merry way