3-22-1: BUG HUNT injector.c manipulations
I've been commenting and uncommenting various code blocks to troubleshoot this very frustrating BUG.  My current thought is that the wrapper around mprotect() isn't working.  It's not working because I'm calling it on the wrong address.  I'm calling it on the wrong address because I'm passing in the Relative Virtual Address of the first r-xp memory section I find.  I'll need an Absolute (?) Virtual Address instead.