3-22-1: BUG HUNT Much Debugging in injector.c
I added some DEBUGGING statements to verify the registers were being captured and modified properly.  Also, some out-of-logic-branch-flow DEBUGGING logic to attempt to verify the new RIP was properly set.  All of that appears to be functional.

I discovered today that mprotect() isn't doing what I want because it's operating on MY process instead of the victim process.

Changed tactics:
	- Get rw-p virt mem segment, modified it, SEGFAULT
	- mmap share mem, copy in shell code, point RIP to the mapped memory, SEGFAULT