memset: Incremental commit of a_m_e.c
automate_memset_experiment.c now includes the functionality to determine all of the input filenames given the combinatorial input macros.  I'm getting a lot of noise from map_mem for missing files.  Next, I'll include some logic to quietly determine if the file exists or not.

Changes to be committed:
	modified:   automate_memset_experiment.c