memset: Incremental commit of a_m_e.c
More functionality.  Biggest update, it now logs missing files to the log_file.md.  Plus, the .md looks good in the github web interface.  Now for the hard part.  Implementing functionality that will parse an ELF to determine whether it calls a specific library function or not.  Oof.

Changes to be committed:
	modified:   automate_memset_experiment.c