3-22-1: BUG TROPHY htrace_write_data()
Resolved the bug in htrace_write_data().  To be honest, I'm not sure *why* it fixed it but I chose a different method of pointer math.  Instead of adding a relative offset to the parameters "source" data, I made a variable for the sole intention of iterating through the source data.  I'm not entirely sure why it worked but it worked.  Leave it be.  My code is now underpinning injector.c.  Clean up commit to follow.

Changes to be committed:
	modified:   Harkletrace.c