fixes: BUG HUNT grandprix synchronization hang
Main thread is hanging on the pipe read.  Not sure why this is happening or what changed to make this happen?  Roll back some commits to find a good state?  Recent issue with pipe2()?  Not sure yet.  This commit is chock-full of DEBUGGING statements.  I know what the problem is but I'm not sure what's causing it.  I did, however, find a note from Past Hark (solid dude, that one).  Maybe I should git blame the note and roll back grandprix.exe to that commit and see what's afoot.

 Changes to be committed:
	modified:   3-18_Threading_Practice-1_Grand_Prix.c
	modified:   Harklepipe.c