3-10-3: Fileroad resolve_symlink() Initial Commit
I prototyped and stubbed out a function, resolve_symlink(), I intend to use to resolve /proc/PID/exe to the absolute path of the executable in question so that I may parse its ELF header.  This is in pursuit of printing the shared objects it was compiled against.
Not yet functional.  Returns NULL.  Added notes inside function definition on how to go after this function definition.

Changes to be committed:
	modified:   Fileroad.c
	modified:   Fileroad.h